+++ src/tasks/complex.rs
@@ -1,6 +1,11 @@
 use crate::{
-    config::Config, environment::RawEnvironment, errors::DevrcResult, interpreter::InterpreterKind,
-    scope::Scope, variables::RawVariables, workshop::Designer,
+    config::Config,
+    environment::RawEnvironment,
+    errors::DevrcResult,
+    interpreter::InterpreterKind,
+    scope::{child_scope, Scope},
+    variables::RawVariables,
+    workshop::Designer,
 };
 use std::cell::RefCell;
 
@@ -10,7 +15,8 @@ use super::{
     arguments::TaskArguments,
     exec::ExecKind,
     params::{ParamValue, Params},
-    result::TaskResult, subtask_call::SubtaskCall,
+    result::TaskResult,
+    subtask_call::SubtaskCall,
 };
 use std::rc::Rc;
 
@@ -86,26 +92,31 @@ impl ComplexCommand {
         config: &Config,
         designer: &Designer,
     ) -> DevrcResult<TaskResult> {
-        let mut local_scope = self.process_variables(parent_scope, args)?;
-
+        let mut local_scope = self.compute_execution_scope(parent_scope, args)?;
         let interpreter = self.get_interpreter(config);
 
         // TODO: register output as variable
-        self.exec.execute(&mut local_scope, config, &interpreter, designer)?;
+        self.exec
+            .execute(&mut local_scope, config, &interpreter, designer)?;
 
         Ok(TaskResult::new())
     }
 
-    /// Prepare template scope
-    pub fn process_variables(
+    pub fn compute_execution_scope(
+        &self,
+        scope_ref: Rc<RefCell<Scope>>,
+        _args: &TaskArguments,
+    ) -> DevrcResult<Scope> {
+        let binding = (*scope_ref).borrow();
+        binding.compute_execution_scope()
+    }
+
+    pub fn get_scope(
         &self,
         parent_scope: Rc<RefCell<Scope>>,
         args: &TaskArguments,
     ) -> DevrcResult<Scope> {
-        let mut scope = Scope {
-            parent: Some(Rc::clone(&parent_scope)),
-            ..Default::default()
-        };
+        let mut scope = child_scope(parent_scope, self.name.clone().unwrap_or_default().as_ref());
 
         // TODO: here devrc can ask user input
         for (key, (value, _)) in args {