+++ src/tasks/mod.rs
@@ -25,8 +25,9 @@ pub mod subtask_call;
 
 pub use crate::tasks::{examples::Examples, exec::ExecKind, params::Params};
 
-use self::subtask_call::SubtaskCall;
-use self::{complex::ComplexCommand, params::ParamValue, result::TaskResult};
+use self::{
+    complex::ComplexCommand, params::ParamValue, result::TaskResult, subtask_call::SubtaskCall,
+};
 use crate::tasks::arguments::TaskArguments;
 
 #[derive(Debug, Deserialize, Clone)]
@@ -56,20 +57,15 @@ pub struct Include {
 }
 
 // TODO: put `ComplexCommand` into `Box`
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize, Default)]
 #[serde(untagged)]
 pub enum TaskKind {
+    #[default]
     Empty,
     Command(String),
     ComplexCommand(ComplexCommand),
     Commands(ExecKind),
     Include(Include),
-}
-
-impl Default for TaskKind {
-    fn default() -> Self {
-        TaskKind::Empty
-    }
 }
 
 impl TaskKind {
@@ -123,7 +119,6 @@ impl TaskKind {
             &format!("\n==> Running task: `{:}` ...", &name),
             &designer.banner(),
         );
-
         let result = match self {
             TaskKind::Empty => return Err(DevrcError::NotImplemented),
             TaskKind::Command(command) => {
@@ -150,9 +145,9 @@ impl TaskKind {
         let result = match self {
             TaskKind::Empty => return Err(DevrcError::NotImplemented),
             TaskKind::Command(value) => {
-                ComplexCommand::from(value).process_variables(parent_scope, args)?
+                ComplexCommand::from(value).get_scope(parent_scope, args)?
             }
-            TaskKind::ComplexCommand(value) => value.process_variables(parent_scope, args)?,
+            TaskKind::ComplexCommand(value) => value.get_scope(parent_scope, args)?,
             TaskKind::Commands(_value) => return Err(DevrcError::NotImplemented),
             TaskKind::Include(_value) => {
                 return Err(DevrcError::NotImplemented);
@@ -177,7 +172,7 @@ impl TaskKind {
     pub fn get_subtasks(&self) -> Option<&Vec<SubtaskCall>> {
         if let TaskKind::ComplexCommand(command) = self {
             if !&command.subtasks.is_empty() {
-                return Some(&command.subtasks)
+                return Some(&command.subtasks);
             }
         }
         None