+++ src/tasks/subtask_call.rs
@@ -1,9 +1,6 @@
-
 use serde::Deserialize;
 
 use crate::{environment::RawEnvironment, variables::RawVariables};
-
-
 
 #[derive(Debug, Deserialize, Clone, Default)]
 pub struct SubtaskCall {
@@ -14,5 +11,4 @@ pub struct SubtaskCall {
 
     #[serde(default)]
     pub variables: RawVariables,
-
 }