+++ tests/test_environment.rs
@@ -1,4 +1,7 @@
-use devrc::environment::{EnvFile, EnvFilesWrapper, FileInclude, FileRemote, RawEnvironment};
+use devrc::{
+    env_file::{EnvFilesInclude, LocalFileImport, UrlImport},
+    environment::RawEnvironment,
+};
 use std::fmt::Debug;
 
 use serde::Deserialize;
@@ -26,41 +29,49 @@ fn test_include_environment_files() {
     #[derive(Debug, Deserialize, Clone)]
     pub struct Container {
         #[serde(rename(deserialize = "env_file"))]
-        files: EnvFilesWrapper,
+        files: EnvFilesInclude,
     }
 
     let content: &str = r#"
 env_file:
   - ./.env
   - /path/to/file_1
-  - remote: http://example.com
+  - url: http://example.com
+    checksum: "checksum"
   - file: /path/to/file_2
 "#;
 
     let container: Container = serde_yaml::from_str::<Container>(content).unwrap();
 
-    if let EnvFile::List(val) = container.files.0 {
-        if let EnvFile::Simple(variant) = &val[0] {
+    if let EnvFilesInclude::List(val) = container.files {
+        if let EnvFilesInclude::Simple(variant) = &val[0] {
             assert_eq!(variant.to_str().unwrap(), "./.env".to_string());
         } else {
             unreachable!();
         }
 
-        if let EnvFile::Simple(variant) = &val[1] {
+        if let EnvFilesInclude::Simple(variant) = &val[1] {
             assert_eq!(variant.to_str().unwrap(), "/path/to/file_1".to_string());
         } else {
             unreachable!();
         }
 
-        if let EnvFile::Remote(FileRemote { remote }) = &val[2] {
-            assert_eq!(remote.to_string(), "http://example.com".to_string());
+        if let EnvFilesInclude::Url(UrlImport {
+            url,
+            ignore_errors: _,
+            checksum: _,
+        }) = &val[2]
+        {
+            assert_eq!(url.to_string(), "http://example.com".to_string());
         } else {
             unreachable!();
         }
 
-        if let EnvFile::File(FileInclude {
+        if let EnvFilesInclude::File(LocalFileImport {
             file,
             ignore_errors: _,
+            path_resolve: _,
+            checksum: _,
         }) = &val[3]
         {
             assert_eq!(file.to_str().unwrap(), "/path/to/file_2".to_string());
@@ -75,12 +86,18 @@ env_file:
 #[test]
 fn test_include_remote() {
     let content: &str = r#"
-remote: http://example.com
+url: http://example.com
+checksum: "checksum"
 "#;
-    let container = serde_yaml::from_str::<EnvFile>(content).unwrap();
+    let container = serde_yaml::from_str::<EnvFilesInclude>(content).unwrap();
 
-    if let EnvFile::Remote(FileRemote { remote }) = &container {
-        assert_eq!(remote.to_owned(), "http://example.com".to_string());
+    if let EnvFilesInclude::Url(UrlImport {
+        url,
+        ignore_errors: _,
+        checksum: _,
+    }) = &container
+    {
+        assert_eq!(url.to_owned(), "http://example.com".to_string());
     } else {
         unreachable!()
     }
@@ -91,11 +108,13 @@ fn test_include_file() {
     let content: &str = r#"
 file: /path/to/file_2
 "#;
-    let container = serde_yaml::from_str::<EnvFile>(content).unwrap();
+    let container = serde_yaml::from_str::<EnvFilesInclude>(content).unwrap();
 
-    if let EnvFile::File(FileInclude {
+    if let EnvFilesInclude::File(LocalFileImport {
         file,
         ignore_errors: _,
+        path_resolve: _,
+        checksum: _,
     }) = &container
     {
         assert_eq!(file.to_str().unwrap(), "/path/to/file_2".to_string());
@@ -109,9 +128,9 @@ fn test_include_simple_file() {
     let content: &str = r#"
 ./.env
 "#;
-    let container = serde_yaml::from_str::<EnvFile>(content).unwrap();
+    let container = serde_yaml::from_str::<EnvFilesInclude>(content).unwrap();
 
-    if let EnvFile::Simple(variant) = container {
+    if let EnvFilesInclude::Simple(variant) = container {
         assert_eq!(variant.to_str().unwrap(), "./.env".to_string());
     } else {
         unreachable!();